Split resolver validation doc guards
